+++ src/configs/address-specific-options.tsx
@@ -9,13 +9,14 @@ type TAddressConfig = Record<
     qrConfig?: TAddressQRConfig
     qrDeeplink?: TAddressQRDeeplinkConfig
     betaFeaturesAvailable?: TFeatureName[]
-    claimAppUrl?: string
+    claimAppUrl?: string,
+    claimUrlSchema?: string
   }
 >
 
 const COINBASE_CONFIG = {
   qrConfig: 'coinbase',
-  qrDeeplink: 'https://go.cb-w.com/dapp?cb_url=%URL%'
+  claimUrlSchema: 'https://wallet.coinbase.com/claim?tk=code&k=<CODE>&c=<CHAIN_ID>&v=<VERSION>&src=d'
 }
 
 const LEDGER_CONFIG = {
@@ -35,9 +36,8 @@ export const addressSpecificOptions: TAddressConfig = {
   '0xf2b95635fc6cf3ebc8b5ca01c4e683a22b2e662c': COINBASE_CONFIG,
   '0xf92f4c77ab90b6e0ee2ccb24fa176d151c5ba9d0': ALPHEMY_CONFIG,
   '0x241365e6eed258ac4a2824424b6802c0240e13eb': ALPHEMY_CONFIG,
-  '0xb4c3d57327d4fc9bcc3499963e21db1a5435d537': LEDGER_CONFIG,
   '0x1e5fac5151e7d6db73a0aa7d740e9fd7063d6dbb': ALPHEMY_CONFIG,
   '0x2254c74edd4220883d881894f349b954e15ec7a0': ALPHEMY_CONFIG,
-  '0xadf49b9f133fb137e82b24f06d23e49c51f586c7': LEDGER_CONFIG, 
+  '0xadf49b9f133fb137e82b24f06d23e49c51f586c7': LEDGER_CONFIG,
   '0x0020eedeb84da1491081c1409b684f459b8f0697': COINBASE_CONFIG
 }