+++ src/data/store/reducers/user/async-actions/get-erc20-token-list.tsx
@@ -1,5 +1,3 @@
-import { ethers } from 'ethers'
-import contracts from 'configs/contracts'
 import { prepareERC20Url, alertError } from 'helpers'
 
 const getERC20TokenList = async (